refactor(indexes): add sorted index value and attached to the sorted list index
US-32041
